Move concept config to its own file
